# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neighborhood Commons API",
    "description": "The Neighborhood Commons API: typed substrate for neighborhood-scale public facts. Five Schema.org-aligned primitives (Place, Organization, Event, Broadcast, List), verification anchored on Type A profile authority, constrained-publishing model with three authority paths (entity-runs-it, pipeline-proxies, witnessed-with-evidence). Read free, write through linked organizations. CC BY 4.0.\n\n**This document is the Spec, part of the Commons Contract.** The Contract is three files together: the Spec (this document — machine-readable, authoritative), the Guide (`/llms.txt` — narrative companion), and the Log (`CHANGELOG.md` — dated record of every change). Rule when they disagree: Spec wins. Guide explains. Log dates.\n\n**3.0.0** establishes the standard four-role event provenance frame (`docs/four-roles.md`) and the type-general provenance method doctrine (`docs/provenance.md`). `source.publisher` is removed (the role is filled by `organizer.name`); `source.method` uses the standard vocabulary (`self_asserted` / `proxied` / `witnessed`); `source.url` carries the proxied source URL when applicable. Organizations, Broadcasts, and Lists gain a `method` field — Organizations admit `seeded` for bulk-imported, awaiting-uptake rows. This is the launch contract; from here forward the substrate is additive-only.",
    "version": "3.0.0",
    "license": {
      "name": "CC-BY-4.0",
...
    },
...
  },
  "servers": [
...
  "paths": {
...
    "/service/verifications/path": {
      "get": {
        "summary": "Discover the required verification path",
        "description": "Routing authority. Given a target and identifier, the Commons returns which submission endpoint to call next. Apps must call this first and follow the result — submission endpoints reject mismatches.",
        "operationId": "serviceVerificationPath",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "target_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "organization"
              ]
            },
            "description": "Always `organization`. Retained as a query parameter for forward compatibility; the substrate may admit additional target types in future minor versions."
          },
          {
            "name": "target_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "identifier_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "email"
              ]
            }
          },
          {
            "name": "identifier_value",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Path discovery result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationPathResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Target not found"
          }
        }
      }
    },
    "/service/verifications/challenges": {
      "post": {
        "summary": "Issue a verification challenge (auto-track)",
        "description": "Auto-track verification path: send a one-time code to the identifier (email). Used for business-domain emails on heavy-rigor organizations (commercial=true) and any email on light-rigor organizations (community groups, collectives, etc.). Personal-email domains for heavy-rigor targets are rejected here with WRONG_METHOD — call /service/verifications/manual instead.",
        "operationId": "serviceCreateVerificationChallenge",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationChallengeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created; email sent (with the calling key's brand_config sender identity)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationChallengeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "WRONG_METHOD — this identifier requires the manual review path. Response includes redirect hint."
          }
        }
      }
    },
    "/service/verifications/challenges/{id}/confirm": {
      "post": {
        "summary": "Confirm a verification challenge",
        "description": "Submit the one-time code received via email. On match: an entry is added to the target's verified identifiers. On miss: attempt counter increments; challenge expires after configured limit.",
        "operationId": "serviceConfirmVerificationChallenge",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationConfirmInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification outcome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOutcome"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Challenge not found or expired"
          },
          "409": {
            "description": "Challenge already consumed"
          }
        }
      }
    },
    "/service/verifications/manual": {
      "post": {
        "summary": "Submit manual-review verification evidence",
        "description": "Manual-review path: app submits structured evidence for human review. Required evidence schema is Commons-defined and validated at submit. Apps with `verification_authority` for the matching method auto-approve on submit; others queue for admin review.\n\nFor heavy-rigor org targets, this is the path for personal-email domains. For light-rigor targets, this path is rarely used (challenges suffice).",
        "operationId": "serviceSubmitManualVerification",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationManualInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification outcome — verified (auto-approved by app authority), pending (queued), or rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOutcome"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "WRONG_METHOD — this identifier should use the challenges path."
          }
        }
      }
    },
    "/service/verifications/pending": {
      "get": {
        "summary": "List pending verification reviews",
        "description": "Admin-tier read of the manual review queue. Only callers whose service key has `is_admin=true` (or has `verification_authority` covering the relevant method) see results.",
        "operationId": "serviceListPendingVerifications",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "reviews"
                  ],
                  "properties": {
                    "reviews": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PendingReview"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/service/verifications/pending/{id}/approve": {
      "post": {
        "summary": "Approve a pending verification review",
        "description": "Admin or authorized-app action. Inserts the verified identifier and marks the review approved. Approval criteria documented in `docs/verification-policy.md` — reviewer is bound by Commons-defined floor.",
        "operationId": "serviceApprovePendingVerification",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOutcome"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Already reviewed"
          }
        }
      }
    },
    "/service/verifications/pending/{id}/reject": {
      "post": {
        "summary": "Reject a pending verification review",
        "description": "Admin or authorized-app action. Records the rejection with a Commons-defined reason code. Rejection is a permanent record but doesn't prevent future fresh submissions for the same identifier.",
        "operationId": "serviceRejectPendingVerification",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRejectInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rejected"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Already reviewed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VerificationChallengeInput": {
        "type": "object",
        "description": "Only organizations verify; there is no targetType discriminator.",
        "required": [
          "organizationId",
          "identifierType",
          "identifierValue"
        ],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "identifierType": {
            "type": "string",
            "enum": [
              "email"
            ]
          },
          "identifierValue": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "browseApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional for read endpoints (higher rate limit bucket). Required for webhook management."
      },
      "serviceApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-tier API key. Scoped: can only modify data for accounts linked to this key. Admin keys bypass scoping."
      }
    }
  }
}